% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTeam.R
\name{buildTeam}
\alias{buildTeam}
\title{Build Team}
\usage{
buildTeam(playerName, data = df)
}
\arguments{
\item{playerName}{the name of an active MLB player}

\item{data}{the data source, either pitcherData or hitterData}
}
\description{
creates a dataframe of desired fantasy baseball players
}
\examples{
buildTeam("Ronald Acuna Jr.", hitterData)
}
