% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitterProjections.R
\name{hitterProjections}
\alias{hitterProjections}
\title{Hitter Projections}
\usage{
hitterProjections(hitterName)
}
\arguments{
\item{hitterName}{the name of an active MLB hitter}
}
\description{
calculates total fantasy points for the season for a given hitter
}
\examples{
hitterProjections("Ronald Acuna Jr.")
}
