% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitcherProjections.R
\name{pitcherProjections}
\alias{pitcherProjections}
\title{Pitcher Projections}
\usage{
pitcherProjections(pitcherName)
}
\arguments{
\item{pitcherName}{the name of an active MLB pitcher}
}
\description{
calculates total fantasy points for the season for a given pitcher
}
\examples{
pitcherProjections("Spencer Strider")
}
